pyproject.toml
[tool.poetry]
name = "pz-mod-builder"
version = "0.1.0"
description = "Python 3 tools for building Project Zomboid version b42 mods"
authors = ["pz_mod_builder contributors"]
readme = "README.md"
packages = [{include = "pz_mod_builder", from = "src"}]
license = "MIT"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^10.0.0"
ruff = "^0.1.0"
rich = "^13.0.0"
questionary = "^2.0.0"

[tool.poetry.scripts]
pzmod = "pz_mod_builder.cli:main"

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "I"]
ignore = []

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
